+++ app/src/main/java/org/hyperledger/besu/controller/MergeBesuControllerBuilder.java
@@ -187,7 +187,13 @@ public class MergeBesuControllerBuilder extends BesuControllerBuilder {
         isParallelTxProcessingEnabled,
         balConfiguration,
         metricsSystem,
-        evmConfiguration);
+        evmConfiguration,
+        protocolScheduleCustomization);
+  }
+
+  @Override
+  protected boolean supportsProtocolScheduleCustomization() {
+    return true;
   }
 
   /**